back_end/requirement.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
Django==4.1.7
django-filter==22.1
django-guardian==2.4.0
django-rest-swagger==2.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.4.1
MarkupSafe==2.1.2
openapi-codec==1.3.2
Pillow==9.4.0
psycopg2-binary==2.9.6
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
simplejson==3.18.3
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
